fixed bug

+++ immunization_llms/evaluation.py
@@ -6,7 +6,7 @@ from tqdm import tqdm
 from transformers import pipeline
 import gem_metrics
 
-harmfulness_pipe = pipeline('text-classification', model='/scratch/gpfs/bw1822/nlp_checkpoints/deberta-v3-xsmall-beavertails-harmful-qa-classifier')
+harmfulness_pipe = pipeline('text-classification', model='/scratch/gpfs/bw1822/nlp_checkpoints/deberta-v3-xsmall-beavertails-harmful-qa-classifier', torch_dtype=torch.bfloat16)
 
 def evaluate_harmfulness(
     model,
